ref: Separate downloads from object handling
